info: move brand admin helper functions to dedicated file
